+++ mods/CreateDragonPlus/src/integration/simulated/java/plus/dragons/createdragonsplus/integration/simulated/data/internal/package-info.java
@@ -1,7 +1,7 @@
 @FieldsAreNonnullByDefault
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
-package plus.dragons.createdragonsplus.integration.simulated.common.kinetics.fan;
+package plus.dragons.createdragonsplus.integration.simulated.data.internal;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;